Rename Resource into AuthorizationResource
